simplify the project Management to make it easier to test
