+++ src/Components/ProjectManagement/CodeFeedback/CodeFeedback.js
@@ -160,9 +160,7 @@ const CodeFeedback = props => {
         setDoneLoading(false);
         setAllResponsesGraded(true);
         let response = { data: { message: "success" } };
-        response = await axios.post("/retreiveFeedbackcodes", {
-          fullname
-        });
+        response = await axios.post("researchers/retreiveFeedbackcodes");
         if (response.data.message === "success") {
           setExplanationIdsList(response.data.codeIds);
           setDoneLoading(true);
